bug fixed for csrxchg instruction; add test13_csrxchg
csrxchg is weird. Because the csrwr instruction is already had the
exchange part.

csrwr rd, csr

write rd into csr indexed register, and the old value from the csr
register will be sent back to rd.

csrxchg rd, rj, csr

similiar to csrwr, rj contains the mask that the csr register bit
only can be changed if the corresponding mask bit is 1.

The bug was that csrxchg write the csr register bit 0 if the bit
is not masked (being 0).