interrupt wait until inst_vld_e=1 to notify the core, add intr_hold
There may be bubbles in the pipeline, when interrupt notices the core,
ifu_exu_pc_e may point at a bubble, and the real instrution is already
at _m. So, _e instruction gets executed twice.

The reason of bubbles is that currently fetching an instruction takes 2
cycles due to the axi bus. The pipeline should execute correctly no
matter how many cycles it takes to fetch the instruction.

   // intr wait for inst_vld_e to be 1
   // meaning, only _e has a valid instruction to signal intrrupt to core
   // there may be bubble in _e stage

   // scenario 1:
   //
   // intr_in            : _-______
   // inst_vld_e         : _-______
   //
   // intr_hold_q        : ________
   // intr               : _-______
   //

   // scenario 2:
   //
   // intr_in            : _-______
   // inst_vld_e         : ___-____
   //
   // intr_hold_q        : __--____
   // intr               : ___-____
   //