make inst_req -_-_-_-_, but cant use inst_busy
inst_busy forms a loop with inst_req. although a register in the
middle breaks it, but altera gives warning.

Warning (332125): Found combinational loop of 1541 nodes
...

Due to previous change, inst_req can not always pull high because ar
channel sends two consecutive transactions with the same ID.  That is
illegal if it is not axi interleaving, or the axi matrix does not
support interleaving. Interleaving

Use inst_req and inst_valid_f to make inst_req only high when actually
request a fetch. In current setup, inst_valid_f is signaled in the next
cycle, and that makes inst_req -_-_-_-... For example, if inst_valid_f
is signaled after 2 cycles, then inst_req will be --_--_--_, and that
would be a problem.

Currently, inst_req wont conflict with ld and st. Lsu delay ld's ar
transactione one cycle to avoid it.