Bug Fix: In exu_ecl, intr Not Synchronized to Start of Next Instruction
If an external interrupt occurs during a load/store that stalls the ifu,
the ecl interrupt synchronization mechanism fails. This happens because
when exu_ifu_stall_req is high, ifu_exu_valid_e also stays at 1, leading
the logic to incorrectly interpret this as the start of the next
instruction.

test32_ext_intr_during_ld, has been added to address this issue.